+++ HISTORY.md
@@ -3,3 +3,10 @@
 <https://www.conventionalcommits.org/en/v1.0.0/>
 
 ## [unreleased]
+
+## [0.2.0] - 2025-12-28 | Change sort order
+
+- feat: marker sort order can be cahnged with  `/gps asc` or `/gps desc` [0.2.0]
+- fix: delete button was shown in radial view for global markers (for non admins) [0.1.2]
+
+## [0.1.0] - 2025-12-28 | Initial release